backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic==2.8.2
pydantic-settings==2.4.0
alembic==1.13.2
python-multipart==0.0.9
psycopg[binary]==3.2.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-dotenv==1.0.1
pytest==8.2.2
httpx==0.27.0
